Avoid clashing Environment between Shared and SwiftUI
